// src/app/interfaces/users.ts
export interface Users {
    id?: number;
    cedula?: string;
    nombres?: string;
    apellidos?: string;
    celular?: string;
    correo?: string;
    id_genero?: number;
    pwd?: string;
    fecha_nacimiento?: string;
    id_rol?: number;
    id_personal_registro?: string;
    fecha_registro?: string;
    id_estado?: number;
    fecha_modificacion?: string;
    ultima_actividad?: string;
}